refactor(factory) code cleanup v3
