+++ src/Database/UnitOfWorkFactory.cs
@@ -1,7 +1,7 @@
 ﻿using Qubiz.QuizEngine.Database.Repositories;
 using Qubiz.QuizEngine.Infrastructure;
 
-namespace Qubiz.QuizEngine.Services
+namespace Qubiz.QuizEngine.Database
 {
     public class UnitOfWorkFactory : IUnitOfWorkFactory
     {
